requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==4.2.11
django-simple-history==3.5.0
djangorestframework==3.15.1
docker==7.0.0
exceptiongroup==1.2.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytest==8.2.0
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3
